% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSHIFT_interface.R
\name{load_sample_data}
\alias{load_sample_data}
\title{Load Sample Data}
\usage{
load_sample_data()
}
\value{
Data frame containing sample data.
}
\description{
This function loads the sample data included in the package.
}
